Added & tested isRowMagic and isColMagic methods As well as: - Fixed indentation in Tester.java - Added newlines to separate Tester output using overloaded `print`
I included these two in the same commit because of how simple and
similar the two methods are given my helper methods. As for changing
Tester so much in one commit, they're relatively small changes that I
wanted to do before testing the new methods. In hindsight, I could
have changed Tester and committed that, then tested the new methods,
but I already typed the testing code and didn't want to go through the
trouble of temporarily taking them out for a commit.